requirements.txt
numpy
tensorflow-cpu
google-genai
flask
tensorflowjs
pillow
psycopg2-binary
flask-sqlalchemy
flask-login
flask-mail
requests
python-dotenv
geopy
folium
plotly
dash
dash-bootstrap-components
requests-cache 
retry-requests 
pandas
openmeteo-requests
gunicorn
